Breakup the P3 scheme file
