+++ src/CloudMicrophysics.jl
@@ -15,7 +15,7 @@ include("Microphysics0M.jl")
 include("Microphysics1M.jl")
 include("Microphysics2M.jl")
 include("TerminalVelocity.jl")
-include("P3Scheme.jl")
+include("P3.jl")
 include("MicrophysicsNonEq.jl")
 include("AerosolModel.jl")
 include("AerosolActivation.jl")
