+++ src/P3_particle_properties.jl
@@ -1,0 +1,228 @@
+"""
+    α_va_si(p3)
+
+ - p3 - a struct with P3 scheme parameters
+
+Returns `α_va` coefficient for the assumed particle mass(size) relation for
+large unrimed ice and dense nonspherical ice, in base SI units: kg m^(-β_va).
+`β_va` is another coefficient of the mass(size) relation.
+From measurements of mass grown by vapor diffusion and aggregation
+in midlatitude cirrus by Brown and Francis (1995)
+doi: 10.1175/1520-0426(1995)012<0410:IMOTIW>2.0.CO;2
+"""
+α_va_si(p3::PSP3{FT}) where {FT} = p3.α_va * 10^(6 * p3.β_va - 3)
+
+"""
+    D_th_helper(p3)
+
+ - p3 - a struct with P3 scheme parameters
+
+Returns the critical size separating spherical and nonspherical ice, in meters.
+Eq. 8 in Morrison and Milbrandt (2015).
+"""
+D_th_helper(p3::PSP3{FT}) where {FT} =
+    (FT(π) * p3.ρ_i / 6 / α_va_si(p3))^(1 / (p3.β_va - 3))
+
+"""
+    D_cr_helper(p3, F_r, ρ_g)
+
+ - p3 - a struct with P3 scheme parameters
+ - F_r - rime mass fraction (q_rim/q_i) [-]
+ - ρ_g - is the effective density of a spherical graupel particle [kg/m^3]
+
+Returns the size of equal mass for graupel and partially rimed ice, in meters.
+Eq. 14 in Morrison and Milbrandt (2015).
+"""
+function D_cr_helper(p3::PSP3{FT}, F_r::FT, ρ_g::FT) where {FT}
+    α_va = α_va_si(p3)
+    return (1 / (1 - F_r) * 6 * α_va / FT(π) / ρ_g)^(1 / (3 - p3.β_va))
+end
+
+"""
+    D_gr_helper(p3, ρ_g)
+
+ - p3 - a struct with P3 scheme parameters
+ - ρ_g - is the effective density of a spherical graupel particle [kg/m^3]
+
+Returns the size of equal mass for graupel and unrimed ice, in meters.
+Eq. 15 in Morrison and Milbrandt (2015).
+"""
+function D_gr_helper(p3::PSP3{FT}, ρ_g::FT) where {FT}
+    α_va = α_va_si(p3)
+    return (6 * α_va / FT(π) / ρ_g)^(1 / (3 - p3.β_va))
+end
+
+"""
+    ρ_g_helper(ρ_r, F_r, ρ_d)
+
+ - ρ_r - rime density (q_rim/B_rim) [kg/m^3]
+ - F_r - rime mass fraction (q_rim/q_i) [-]
+ - ρ_g - is the effective density of a spherical graupel particle [kg/m^3]
+
+Returns the density of total (deposition + rime) ice mass for graupel, in kg/m3
+Eq. 16 in Morrison and Milbrandt (2015).
+"""
+ρ_g_helper(ρ_r::FT, F_r::FT, ρ_d::FT) where {FT} = F_r * ρ_r + (1 - F_r) * ρ_d
+
+"""
+    ρ_d_helper(p3, D_cr, D_gr)
+
+ - p3 - a struct with P3 scheme parameters
+ - D_cr - is the size of equal mass for graupel and partially rimed ice, in meters
+ - D_gr - the size of equal mass for graupel and unrimed ice, in meters
+
+Returns the density of unrimed ice mass, in kg/m3
+Eq. 17 in Morrison and Milbrandt (2015).
+"""
+function ρ_d_helper(p3::PSP3{FT}, D_cr::FT, D_gr::FT) where {FT}
+    α_va = α_va_si(p3)
+    β_m2 = p3.β_va - 2
+    return 6 * α_va * (D_cr^β_m2 - D_gr^β_m2) / FT(π) / β_m2 /
+           max(D_cr - D_gr, eps(FT))
+end
+
+"""
+    thresholds(p3, ρ_r, F_r)
+
+ - p3 - a struct with P3 scheme parameters
+ - ρ_r - rime density (q_rim/B_rim) [kg/m^3]
+ - F_r - rime mass fraction (q_rim/q_i) [-]
+
+Solves the nonlinear system consisting of D_cr, D_gr, ρ_g, ρ_d
+for a given rime density and rime mass fraction.
+Returns a named tuple containing:
+ - D_cr - is the threshold size separating partially rimed ice and graupel [m],
+ - D_gr - is the threshold size separating graupel and dense nonspherical ice [m],
+ - ρ_g - is the effective density of a spherical graupel particle [kg/m3],
+ - ρ_d - is the density of the unrimed portion of the particle [kg/m3],
+"""
+function thresholds(p3::PSP3{FT}, ρ_r::FT, F_r::FT) where {FT}
+
+    @assert F_r >= FT(0)   # rime mass fraction must be positive ...
+    @assert F_r < FT(1)    # ... and there must always be some unrimed part
+
+    if F_r == FT(0)
+        return (; D_cr = FT(0), D_gr = FT(0), ρ_g = FT(0), ρ_d = FT(0))
+    else
+        @assert ρ_r > FT(0)   # rime density must be positive ...
+        @assert ρ_r <= p3.ρ_l # ... and as a bulk ice density can't exceed the density of water
+
+        P3_problem(ρ_d) =
+            ρ_d - ρ_d_helper(
+                p3,
+                D_cr_helper(p3, F_r, ρ_g_helper(ρ_r, F_r, ρ_d)),
+                D_gr_helper(p3, ρ_g_helper(ρ_r, F_r, ρ_d)),
+            )
+
+        ρ_d =
+            RS.find_zero(
+                P3_problem,
+                RS.SecantMethod(FT(0), FT(1000)),
+                RS.CompactSolution(),
+            ).root
+        ρ_g = ρ_g_helper(ρ_r, F_r, ρ_d)
+
+        return (;
+            D_cr = D_cr_helper(p3, F_r, ρ_g),
+            D_gr = D_gr_helper(p3, ρ_g),
+            ρ_g,
+            ρ_d,
+        )
+    end
+end
+
+"""
+    mass_(p3, D, ρ, F_r)
+
+ - p3 - a struct with P3 scheme parameters
+ - D - maximum particle dimension [m]
+ - ρ - bulk ice density (ρ_i for small ice, ρ_g for graupel) [kg/m3]
+ - F_r - rime mass fraction [q_rim/q_i]
+
+Returns mass as a function of size for differen particle regimes [kg]
+"""
+# for spherical ice (small ice or completely rimed ice)
+mass_s(D::FT, ρ::FT) where {FT <: Real} = FT(π) / 6 * ρ * D^3
+# for large nonspherical ice (used for unrimed and dense types)
+mass_nl(p3::PSP3, D::FT) where {FT <: Real} = α_va_si(p3) * D^p3.β_va
+# for partially rimed ice
+mass_r(p3::PSP3, D::FT, F_r::FT) where {FT <: Real} =
+    α_va_si(p3) / (1 - F_r) * D^p3.β_va
+
+"""
+    p3_mass(p3, D, F_r, th)
+
+ - p3 - a struct with P3 scheme parameters
+ - D - maximum particle dimension
+ - F_r - rime mass fraction (q_rim/q_i)
+ - th - P3Scheme nonlinear solve output tuple (D_cr, D_gr, ρ_g, ρ_d)
+
+Returns mass(D) regime, used to create figures for the docs page.
+"""
+function p3_mass(
+    p3::PSP3,
+    D::FT,
+    F_r::FT,
+    th = (; D_cr = FT(0), D_gr = FT(0), ρ_g = FT(0), ρ_d = FT(0)),
+) where {FT <: Real}
+    D_th = D_th_helper(p3)
+    if D_th > D
+        return mass_s(D, p3.ρ_i)          # small spherical ice
+    elseif F_r == 0
+        return mass_nl(p3, D)             # large nonspherical unrimed ice
+    elseif th.D_gr > D >= D_th
+        return mass_nl(p3, D)             # dense nonspherical ice
+    elseif th.D_cr > D >= th.D_gr
+        return mass_s(D, th.ρ_g)          # graupel
+    elseif D >= th.D_cr
+        return mass_r(p3, D, F_r)         # partially rimed ice
+    end
+end
+
+"""
+    A_(p3, D)
+
+ - p3 - a struct with P3 scheme parameters
+ - D - maximum particle dimension
+
+Returns particle projected area as a function of size for different particle regimes
+"""
+# for spherical particles
+A_s(D::FT) where {FT <: Real} = FT(π) / 4 * D^2
+# for nonspherical particles
+A_ns(p3::PSP3, D::FT) where {FT <: Real} = p3.γ * D^p3.σ
+# partially rimed ice
+A_r(p3::PSP3, F_r::FT, D::FT) where {FT <: Real} =
+    F_r * A_s(D) + (1 - F_r) * A_ns(p3, D)
+
+"""
+    p3_area(p3, D, F_r, th)
+
+ - p3 - a struct with P3 scheme parameters
+ - D - maximum particle dimension
+ - F_r - rime mass fraction (q_rim/q_i)
+ - th - P3Scheme nonlinear solve output tuple (D_cr, D_gr, ρ_g, ρ_d)
+
+Returns area(D), used to create figures for the documentation.
+"""
+function p3_area(
+    p3::PSP3,
+    D::FT,
+    F_r::FT,
+    th = (; D_cr = FT(0), D_gr = FT(0), ρ_g = FT(0), ρ_d = FT(0)),
+) where {FT <: Real}
+    # Area regime:
+    if D_th_helper(p3) > D
+        return A_s(D)                      # small spherical ice
+    elseif F_r == 0
+        return A_ns(p3, D)                 # large nonspherical unrimed ice
+    elseif th.D_gr > D >= D_th_helper(p3)
+        return A_ns(p3, D)                 # dense nonspherical ice
+    elseif th.D_cr > D >= th.D_gr
+        return A_s(D)                      # graupel
+    elseif D >= th.D_cr
+        return A_r(p3, F_r, D)             # partially rimed ice
+    else
+        throw("D not in range")
+    end
+end
